Split test_session_skills_codex.py into 3 per-concern files (#4611)
Mechanical refactor per plan - no production code changes, no test assertion
changes. Each test continues to execute from its new home and the source
file is deleted.

- test_session_skills_codex_layout.py: init-session, generated-home,
  profile-skill, managed-orchestrator, materialization delegation, and
  projection safety tests.
- test_session_skills_codex_locking.py: lease acquisition, contention,
  rollback, symlink-guard, and unowned-cleanup tests.
- test_session_skills_codex_persistent_root.py: TestResolvePersistentSessionRoot
  + TestResolvePersistentSessionRoots.

The exception classes (_BodyFailure, _DeletionFailure, _ReleaseFailure),
helpers (_catalog_context, _materialize, _managed, _make_codex_backend,
_stub_backend), and the codex_env fixture are moved to _helpers.py /
conftest.py to avoid cross-test-file imports (pyright rejects cross-file
imports under tests/, and the project convention is to centralize shared
test utilities in _helpers.py).

The layer enforcement allowlist is updated to point the new files at
'autoskillit.execution' (replacing the old test_session_skills_codex.py
entry). The old test_skills.py 'autoskillit.config' entry is migrated to
test_skills_categories.py.

Closes #4611
